+++ packages/react/src/components/Swap/SwapForm.tsx
@@ -1,4 +1,5 @@
 import {
+  Avatar,
   Box,
   Button,
   Center,
@@ -13,25 +14,30 @@ import {
   Link,
   Menu,
   MenuButton,
+  MenuItem,
+  MenuList,
   ScaleFade,
   Text,
   Tooltip,
   VStack,
 } from "@chakra-ui/react";
+import { Spinner } from "../Spinner";
 import { yupResolver } from "@hookform/resolvers/yup";
 import { useWallet } from "@solana/wallet-adapter-react";
 import { PublicKey } from "@solana/web3.js";
 import {
-  IPricingCurve,
   ITokenBonding,
   SplTokenBonding,
 } from "@strata-foundation/spl-token-bonding";
 import { BondingPricing } from "@strata-foundation/spl-token-bonding/dist/lib/pricing";
 import React, { useEffect, useRef, useState } from "react";
 import { useForm } from "react-hook-form";
+import { BsChevronDown } from "react-icons/bs";
 import { RiArrowUpDownFill, RiInformationLine } from "react-icons/ri";
 import * as yup from "yup";
-import { useFtxPayLink, useProvider } from "../../hooks";
+import { useFtxPayLink, useProvider, useTokenMetadata } from "../../hooks";
+import { Royalties } from "./Royalties";
+import { TransactionInfo, TransactionInfoArgs } from "./TransactionInfo";
 
 export interface ISwapFormValues {
   topAmount: number;
@@ -47,48 +53,70 @@ const validationSchema = yup
   })
   .required();
 
-const humanReadablePercentage = (u32: number) => {
-  if (u32 && u32 !== 0) {
-    return ((u32 / 4294967295) * 100).toFixed(2);
-  }
-  return 0;
-};
-
 export interface ISwapFormProps {
-  action: "buy" | "sell";
   isSubmitting: boolean;
   onConnectWallet: () => void;
-  onFlipTokens: (tokenBonding: PublicKey, action: "buy" | "sell") => void;
+  onTradingMintsChange: (args: { base: PublicKey; target: PublicKey }) => void;
   onBuyBase: (tokenBonding: PublicKey) => void;
   onSubmit: (values: ISwapFormValues) => Promise<void>;
-  tokenBonding: ITokenBonding;
-  pricing: BondingPricing;
-  base: {
-    name: string;
-    ticker: string;
-    icon: React.ReactNode;
-    publicKey: PublicKey;
-  };
-  target: {
-    name: string;
-    ticker: string;
-    icon: React.ReactNode;
-    publicKey: PublicKey;
-  };
-  ownedBase: number;
+  tokenBonding: ITokenBonding | undefined;
+  pricing: BondingPricing | undefined;
+  baseOptions: PublicKey[];
+  targetOptions: PublicKey[];
+  base:
+    | {
+        name: string;
+        ticker: string;
+        image: string | undefined;
+        publicKey: PublicKey;
+      }
+    | undefined;
+  target:
+    | {
+        name: string;
+        ticker: string;
+        image: string | undefined;
+        publicKey: PublicKey;
+      }
+    | undefined;
+  ownedBase: number | undefined;
   spendCap: number;
   feeAmount?: number;
+  extraTransactionInfo?: Omit<TransactionInfoArgs, "formRef">[];
 }
 
 function roundToDecimals(num: number, decimals: number): number {
   return Math.trunc(num * Math.pow(10, decimals)) / Math.pow(10, decimals);
 }
 
+function MintMenuItem({
+  mint,
+  onClick,
+}: {
+  mint: PublicKey;
+  onClick: () => void;
+}) {
+  const { image, metadata } = useTokenMetadata(mint);
+
+  return (
+    <MenuItem
+      onClick={onClick}
+      icon={
+        <Center w={8} h={8} color="white" bg="indigo.500" rounded="full">
+          <Avatar w={"100%"} h={"100%"} size="sm" src={image} />
+        </Center>
+      }
+    >
+      <Text>{metadata?.data.symbol}</Text>
+    </MenuItem>
+  );
+}
+
 export const SwapForm = ({
-  action,
+  extraTransactionInfo,
   isSubmitting,
   onConnectWallet,
-  onFlipTokens,
+  onTradingMintsChange,
   onBuyBase,
   onSubmit,
   tokenBonding,
@@ -98,6 +126,8 @@ export const SwapForm = ({
   ownedBase,
   spendCap,
   feeAmount,
+  baseOptions,
+  targetOptions,
 }: ISwapFormProps) => {
   const formRef = useRef() as React.MutableRefObject<HTMLInputElement>;
   const { connected } = useWallet();
@@ -120,21 +150,34 @@ export const SwapForm = ({
     },
     resolver: yupResolver(validationSchema),
   });
-
-  const isBaseSol = base.publicKey.equals(SplTokenBonding.WRAPPED_SOL_MINT);
-  const isBuying = action === "buy";
+  const isBaseSol = base?.publicKey.equals(SplTokenBonding.WRAPPED_SOL_MINT);
   const topAmount = watch("topAmount");
   const slippage = watch("slippage");
-  const hasBaseAmount = ownedBase >= +(topAmount || 0);
+  const hasBaseAmount = (ownedBase || 0) >= +(topAmount || 0);
   const moreThanSpendCap = +(topAmount || 0) > spendCap;
+
+  const lowMint =
+    base &&
+    target &&
+    pricing?.hierarchy.lowest(base.publicKey, target.publicKey);
+  const isBuying = lowMint && lowMint.equals(target?.publicKey);
 
   const handleConnectWallet = () => onConnectWallet();
 
   const handleUseMax = () =>
-    setValue("topAmount", ownedBase >= spendCap ? spendCap : ownedBase);
+    setValue(
+      "topAmount",
+      (ownedBase || 0) >= spendCap ? spendCap : ownedBase || 0
+    );
 
   const handleFlipTokens = () => {
-    onFlipTokens(tokenBonding.publicKey, isBuying ? "sell" : "buy");
+    if (base && target) {
+      onTradingMintsChange({
+        base: target.publicKey,
+        target: base.publicKey,
+      });
+    }
+
     reset();
   };
 
@@ -142,7 +185,7 @@ export const SwapForm = ({
     if (isBaseSol) {
       window.open(ftxPayLink);
     } else {
-      onBuyBase(tokenBonding.publicKey);
+      onBuyBase(tokenBonding!.publicKey);
     }
   };
 
@@ -152,19 +195,18 @@ export const SwapForm = ({
   };
 
   useEffect(() => {
-    if (topAmount && topAmount >= 0 && tokenBonding && pricing) {
-      if (isBuying) {
-        const buyMax = pricing.buyWithBaseAmount(+topAmount);
+    if (
+      topAmount &&
+      topAmount >= 0 &&
+      tokenBonding &&
+      pricing &&
+      base &&
+      target
+    ) {
+      const amount = pricing.swap(+topAmount, base.publicKey, target.publicKey);
 
-        setValue("bottomAmount", roundToDecimals(buyMax, 9));
-        setRate(`${roundToDecimals(buyMax / topAmount, 9)}`);
-      } else {
-        const sellMax = pricing.sellTargetAmount(+topAmount);
-
-        setValue("bottomAmount", roundToDecimals(sellMax, 9));
-        setRate(`${roundToDecimals(sellMax / topAmount, 9)}`);
-      }
-
+      setValue("bottomAmount", topAmount == 0 ? 0 : roundToDecimals(amount, 9));
+      setRate(`${roundToDecimals(amount / topAmount, 9)}`);
       setFee(`${feeAmount}`);
     } else {
       reset({ slippage: slippage });
@@ -181,6 +223,10 @@ export const SwapForm = ({
     slippage,
   ]);
 
+  if (!base || !target || (connected && (!ownedBase || !pricing))) {
+    return <Spinner />;
+  }
+
   return (
     <Box ref={formRef} w="full">
       <form onSubmit={handleSubmit(handleSwap)}>
@@ -190,11 +236,13 @@ export const SwapForm = ({
               <Text color="gray.600" fontSize="xs">
                 You Pay
               </Text>
-              <Link color="indigo.500" fontSize="xs" onClick={handleBuyBase}>
-                Buy More {base.ticker}
-              </Link>
+              {base && (
+                <Link color="indigo.500" fontSize="xs" onClick={handleBuyBase}>
+                  Buy More {base.ticker}
+                </Link>
+              )}
             </Flex>
-            <InputGroup size="lg">
+            <InputGroup zIndex={100} size="lg">
               <Input
                 isInvalid={!!errors.topAmount}
                 isDisabled={!connected}
@@ -218,8 +266,10 @@ export const SwapForm = ({
                 {connected && (
                   <Menu>
                     <MenuButton
+                      cursor="pointer"
                       isDisabled={!connected}
                       as={Button}
+                      rightIcon={<BsChevronDown />}
                       leftIcon={
                         <Center
                           w={8}
@@ -228,16 +278,33 @@ export const SwapForm = ({
                           bg="indigo.500"
                           rounded="full"
                         >
-                          {base.icon}
+                          <Avatar src={base.image} w="100%" h="100%" />
                         </Center>
                       }
                       borderRadius="20px 6px 6px 20px"
                       paddingX={1.5}
                       bgColor="gray.200"
-                      _hover={{ cursor: "default" }}
                     >
                       {base.ticker}
                     </MenuButton>
+                    <MenuList borderColor="gray.300">
+                      {baseOptions.map((mint) => (
+                        <MintMenuItem
+                          mint={mint}
+                          key={mint.toBase58()}
+                          onClick={() =>
+                            onTradingMintsChange({
+                              base: mint,
+                              target:
+                                target.publicKey &&
+                                mint.equals(target.publicKey)
+                                  ? base.publicKey
+                                  : target.publicKey,
+                            })
+                          }
+                        />
+                      ))}
+                    </MenuList>
                   </Menu>
                 )}
               </InputRightElement>
@@ -268,15 +335,15 @@ export const SwapForm = ({
                   variant="ghost"
                   onClick={handleUseMax}
                 >
-                  Use Max ({ownedBase > spendCap ? spendCap : ownedBase}{" "}
+                  Use Max (
+                  {(ownedBase || 0) > spendCap ? spendCap : ownedBase || 0}{" "}
                   {base.ticker})
                 </Button>
               )}
             </Flex>
             <Divider color="gray.200" />
-            {/* flipping to wum (Selling wum to sol) is disabled in beta*/}
             <IconButton
-              isDisabled={!connected || isBaseSol}
+              isDisabled={!connected}
               aria-label="Flip Tokens"
               size="sm"
               colorScheme="gray"
@@ -291,7 +358,7 @@ export const SwapForm = ({
             <Text color="gray.600" fontSize="xs">
               You Receive
             </Text>
-            <InputGroup size="lg">
+            <InputGroup zIndex={99} size="lg">
               <Input
                 isInvalid={!!errors.bottomAmount}
                 isReadOnly
@@ -318,6 +385,7 @@ export const SwapForm = ({
                 {connected && (
                   <Menu>
                     <MenuButton
+                      rightIcon={<BsChevronDown />}
                       isDisabled={!connected}
                       as={Button}
                       leftIcon={
@@ -328,16 +396,32 @@ export const SwapForm = ({
                           bg="indigo.500"
                           rounded="full"
                         >
-                          {target.icon}
+                          <Avatar src={target.image} w="100%" h="100%" />
                         </Center>
                       }
                       borderRadius="20px 6px 6px 20px"
                       paddingX={1.5}
                       bgColor="gray.200"
-                      _hover={{ cursor: "default" }}
                     >
                       {target.ticker}
                     </MenuButton>
+                    <MenuList borderColor="gray.300">
+                      {targetOptions.map((mint) => (
+                        <MintMenuItem
+                          mint={mint}
+                          key={mint.toBase58()}
+                          onClick={() =>
+                            onTradingMintsChange({
+                              target: mint,
+                              base:
+                                base.publicKey && mint.equals(base.publicKey)
+                                  ? target.publicKey
+                                  : base.publicKey,
+                            })
+                          }
+                        />
+                      ))}
+                    </MenuList>
                   </Menu>
                 )}
               </InputRightElement>
@@ -399,6 +483,7 @@ export const SwapForm = ({
                   {...register("slippage")}
                 />
                 <InputRightElement
+                  zIndex={0}
                   w={4}
                   justifyContent="end"
                   paddingRight={1.5}
@@ -409,65 +494,23 @@ export const SwapForm = ({
               </InputGroup>
             </Flex>
             <Flex justify="space-between" alignItems="center">
-              <Text>Estimated Fees</Text>
+              <Text>Solana Network Fees</Text>
               <Flex>{fee}</Flex>
             </Flex>
-            <Flex justify="space-between" alignItems="center">
-              <HStack>
-                <Text>{isBuying ? base.ticker : target.ticker} Royalties</Text>
-                <Tooltip
-                  isDisabled={!connected}
-                  placement="top"
-                  label={`A purchase fee in ${base.ticker} that is split amongst stakers of ${target.ticker}`}
-                  portalProps={{ containerRef: formRef }}
-                >
-                  <Flex>
-                    <Icon
-                      w={5}
-                      h={5}
-                      as={RiInformationLine}
-                      _hover={{ color: "indigo.500", cursor: "pointer" }}
-                    />
-                  </Flex>
-                </Tooltip>
-              </HStack>
-              <Flex>
-                {humanReadablePercentage(
-                  isBuying
-                    ? tokenBonding.buyBaseRoyaltyPercentage
-                    : tokenBonding.sellBaseRoyaltyPercentage
-                )}
-                %
-              </Flex>
-            </Flex>
-            <Flex justify="space-between" alignItems="center">
-              <HStack>
-                <Text>{isBuying ? target.ticker : base.ticker} Royalties</Text>
-                <Tooltip
-                  isDisabled={!connected}
-                  placement="top"
-                  label={`A percentage of every ${target.ticker} token minted goes to the person who has claimed this token`}
-                  portalProps={{ containerRef: formRef }}
-                >
-                  <Flex>
-                    <Icon
-                      w={5}
-                      h={5}
-                      as={RiInformationLine}
-                      _hover={{ color: "indigo.500", cursor: "pointer" }}
-                    />
-                  </Flex>
-                </Tooltip>
-              </HStack>
-              <Flex>
-                {humanReadablePercentage(
-                  isBuying
-                    ? tokenBonding.buyTargetRoyaltyPercentage
-                    : tokenBonding.sellTargetRoyaltyPercentage
-                )}
-                %
-              </Flex>
-            </Flex>
+            {base &&
+              target &&
+              pricing?.hierarchy
+                .path(base.publicKey, target.publicKey)
+                .map((h) => (
+                  <Royalties
+                    formRef={formRef}
+                    tokenBonding={h.tokenBonding}
+                    isBuying={!!isBuying}
+                  />
+                ))}
+            {(extraTransactionInfo || []).map((i) => (
+              <TransactionInfo formRef={formRef} {...i} key={i.name} />
+            ))}
           </VStack>
           <Box position="relative">
             <ScaleFade
@@ -512,13 +555,7 @@ export const SwapForm = ({
               size="lg"
               type="submit"
               isLoading={awaitingApproval || isSubmitting}
-              loadingText={
-                awaitingApproval
-                  ? "Awaiting Approval"
-                  : action === "buy"
-                  ? "Buying"
-                  : "Selling"
-              }
+              loadingText={awaitingApproval ? "Awaiting Approval" : "Swapping"}
             >
               Trade
             </Button>