+++ packages/spl-token-bonding/src/index.ts
@@ -4,6 +4,7 @@ import {
   createMintInstructions,
   getMintInfo,
   getTokenAccount,
+  sleep,
 } from "@project-serum/common";
 import {
   AccountInfo,
@@ -13,10 +14,13 @@ import {
   NATIVE_MINT,
   Token,
   TOKEN_PROGRAM_ID,
+  u64,
 } from "@solana/spl-token";
 import {
+  Commitment,
   Keypair,
   PublicKey,
+  Signer,
   SystemProgram,
   SYSVAR_CLOCK_PUBKEY,
   SYSVAR_RENT_PUBKEY,
@@ -26,12 +30,14 @@ import {
   AnchorSdk,
   createMetadata,
   Data,
+  getAssociatedAccountBalance,
   InstructionResult,
   percent,
   TypedAccountParser,
 } from "@strata-foundation/spl-utils";
 import BN from "bn.js";
-import { amountAsNum, asDecimal, fromCurve, IPricingCurve } from "./curves";
+import { BondingHierarchy } from "./bondingHierarchy";
+import { fromCurve, IPricingCurve } from "./curves";
 import {
   CurveV0,
   ProgramStateV0,
@@ -39,10 +45,13 @@ import {
   TokenBondingV0,
 } from "./generated/spl-token-bonding";
 import { BondingPricing } from "./pricing";
+import { asDecimal, toBN, toNumber, toU128 } from "./utils";
 
+export * from "./bondingHierarchy";
 export * from "./curves";
-export * from "./pricing";
 export * from "./generated/spl-token-bonding";
+export * from "./pricing";
+export * from "./utils";
 
 /**
  * The curve config required by the smart contract is unwieldy, implementors of `CurveConfig` wrap the interface
@@ -53,50 +62,6 @@ interface ICurveConfig {
 
 interface IPrimitiveCurve {
   toRawPrimitiveConfig(): any;
-}
-
-/**
- * Convert a number to a string avoiding scientific notation
- * @param n
- * @returns
- */
-function toFixedSpecial(num: number, n: number): string {
-  var str = num.toFixed(n);
-  if (str.indexOf("e+") === -1) return str;
-
-  // if number is in scientific notation, pick (b)ase and (p)ower
-  str = str
-    .replace(".", "")
-    .split("e+")
-    .reduce(function (b: any, p: any) {
-      // @ts-ignore
-      return b + Array(p - b.length + 2).join(0);
-    });
-
-  if (n > 0) {
-    // @ts-ignore
-    str += "." + Array(n + 1).join(0);
-  }
-
-  return str;
-}
-
-/**
- * Convert a number to a 12 decimal fixed precision u128
- *
- * @param num Number to convert to a 12 decimal fixed precision BN
- * @returns
- */
-export function toU128(num: number | BN): BN {
-  if (BN.isBN(num)) {
-    return num;
-  }
-
-  if (num == Infinity) {
-    return new BN(0);
-  }
-
-  return new BN(toFixedSpecial(num, 12).replace(".", ""));
 }
 
 /**
@@ -371,7 +336,7 @@ export interface IBuyArgs {
   slippage: number;
 }
 
-export interface IBuyWithBaseArgs {
+export interface ISwapArgs {
   baseMint: PublicKey;
   targetMint: PublicKey;
   /** The payer to run this transaction, defaults to provider.wallet */
@@ -382,6 +347,12 @@ export interface IBuyWithBaseArgs {
   baseAmount: BN | number;
   /** The slippage PER TRANSACTION */
   slippage: number;
+  /** Optionally inject extra instructions before each trade. Usefull for adding txn fees */
+  extraInstructions?: (args: {
+    tokenBonding: ITokenBonding;
+    isBuy: boolean;
+    amount: BN;
+  }) => Promise<InstructionResult<null>>;
 }
 
 export interface ISellArgs {
@@ -413,22 +384,6 @@ export interface ISellBondingWrappedSolArgs {
   owner?: PublicKey /** The owner of the twSOL source account. **Default:** provider wallet */;
   payer?: PublicKey;
   all?: boolean /** Sell all and close this account? **Default:** false */;
-}
-
-function toNumber(numberOrBn: BN | number, mint: MintInfo): number {
-  if (BN.isBN(numberOrBn)) {
-    return amountAsNum(numberOrBn, mint);
-  } else {
-    return numberOrBn;
-  }
-}
-
-function toBN(numberOrBn: BN | number, mint: MintInfo): BN {
-  if (BN.isBN(numberOrBn)) {
-    return numberOrBn;
-  } else {
-    return new BN(Math.ceil(Number(numberOrBn) * Math.pow(10, mint.decimals)));
-  }
 }
 
 /**
@@ -1533,59 +1488,174 @@ export class SplTokenBonding extends AnchorSdk<SplTokenBondingIDL> {
     await this.execute(this.buyInstructions(args), args.payer);
   }
 
-  async buyWithBase({
+  async getTokenAccountBalance(
+    account: PublicKey,
+    commitment: Commitment = "confirmed"
+  ): Promise<BN> {
+    const acct = await this.provider.connection.getAccountInfo(
+      account,
+      commitment
+    );
+    if (acct) {
+      return u64.fromBuffer(AccountLayout.decode(acct.data).amount);
+    }
+
+    return new BN(0);
+  }
+
+  /**
+   * Swap from any base mint to any target mint that are both on a shared link of bonding curves.
+   * Intelligently traverses using either buy or sell, executing multiple txns to either sell baseAmount
+   * or buy with baseAmount
+   *
+   * @param param0
+   */
+  async swap({
     payer = this.wallet.publicKey,
     sourceAuthority = this.wallet.publicKey,
     baseMint,
     targetMint,
     baseAmount,
     slippage,
-  }: IBuyWithBaseArgs): Promise<{ targetAmount: number }> {
-    const hierarchy = await this.getBondingHierarchy(
+    extraInstructions = () =>
+      Promise.resolve({
+        instructions: [],
+        signers: [],
+        output: null,
+      }),
+  }: ISwapArgs): Promise<{ targetAmount: number }> {
+    const hierarchyFromTarget = await this.getBondingHierarchy(
       (
         await SplTokenBonding.tokenBondingKey(targetMint)
       )[0],
       baseMint
     );
-    let current = hierarchy;
-    let highestParent = hierarchy;
-    while (current) {
-      highestParent = current;
-      current = current?.parent;
+    const hierarchyFromBase = await this.getBondingHierarchy(
+      (
+        await SplTokenBonding.tokenBondingKey(baseMint)
+      )[0],
+      targetMint
+    );
+    const hierarchy = [hierarchyFromTarget, hierarchyFromBase].find(
+      (hierarchy) => hierarchy?.contains(baseMint, targetMint)
+    );
+    if (!hierarchy) {
+      throw new Error(
+        `No bonding curve hierarchies found for base or target that contain both ${baseMint.toBase58()} and ${targetMint.toBase58()}`
+      );
     }
+    const isBuy = hierarchy.tokenBonding.targetMint.equals(targetMint);
+    const arrHierarchy = hierarchy?.toArray() || [];
 
-    current = highestParent;
-    while (current) {
-      const tokenBonding = current.tokenBonding;
+    const baseMintInfo = await getMintInfo(this.provider, baseMint);
+
+    let currAmount = toBN(baseAmount, baseMintInfo);
+    for (const subHierarchy of isBuy ? arrHierarchy.reverse() : arrHierarchy) {
+      const tokenBonding = subHierarchy.tokenBonding;
+      const baseIsSol = tokenBonding.baseMint.equals(
+        SplTokenBonding.WRAPPED_SOL_MINT
+      );
       const ata = await Token.getAssociatedTokenAddress(
         ASSOCIATED_TOKEN_PROGRAM_ID,
         TOKEN_PROGRAM_ID,
-        tokenBonding.targetMint,
+        isBuy ? tokenBonding.targetMint : tokenBonding.baseMint,
         sourceAuthority
       );
-      const preBalance =
-        (await this.accountExists(ata)) &&
-        (await (
-          await getTokenAccount(this.provider, ata)
-        ).amount);
-      await this.buy({
-        payer,
-        sourceAuthority,
-        baseAmount,
-        tokenBonding: tokenBonding.publicKey,
-        slippage,
-      });
-      const postBalance = await (
-        await getTokenAccount(this.provider, ata)
-      ).amount;
-      baseAmount = postBalance.sub(preBalance || new BN(0));
 
-      current = current.child;
+      const getBalance = async (): Promise<BN> => {
+        if (!isBuy && baseIsSol) {
+          return new BN(
+            (
+              await this.provider.connection.getAccountInfo(
+                sourceAuthority,
+                "single"
+              )
+            )?.lamports || 0
+          );
+        } else {
+          return this.getTokenAccountBalance(ata, "single");
+        }
+      };
+      const preBalance = await getBalance();
+
+      let instructions: TransactionInstruction[];
+      let signers: Signer[];
+      if (isBuy) {
+        console.log(
+          `Actually doing ${tokenBonding.baseMint.toBase58()} to ${tokenBonding.targetMint.toBase58()}`
+        );
+        ({ instructions, signers } = await this.buyInstructions({
+          payer,
+          sourceAuthority,
+          baseAmount: currAmount,
+          tokenBonding: tokenBonding.publicKey,
+          slippage,
+        }));
+      } else {
+        console.log(
+          `SELL doing ${tokenBonding.baseMint.toBase58()} to ${tokenBonding.targetMint.toBase58()}`
+        );
+        ({ instructions, signers } = await this.sellInstructions({
+          payer,
+          sourceAuthority,
+          targetAmount: currAmount,
+          tokenBonding: tokenBonding.publicKey,
+          slippage,
+        }));
+      }
+
+      const { instructions: extraInstrs, signers: extaSigners } =
+        await extraInstructions({
+          tokenBonding,
+          amount: currAmount,
+          isBuy,
+        });
+      await this.sendInstructions(
+        [...instructions, ...extraInstrs],
+        [...signers, ...extaSigners],
+        payer
+      );
+
+      async function newBalance(tries: number = 0): Promise<BN> {
+        if (tries >= 4) {
+          return new BN(0);
+        }
+        let postBalance = await getBalance();
+        // Sometimes it can take a bit for Solana to catch up
+        // Wait and see if the balance truly hasn't changed.
+        if (postBalance.eq(preBalance)) {
+          console.log(
+            "No balance change detected while swapping, trying again",
+            tries
+          );
+          await sleep(5000);
+          return newBalance(tries + 1);
+        }
+
+        return postBalance;
+      }
+
+      const postBalance = await newBalance();
+
+      currAmount = postBalance!.sub(preBalance || new BN(0));
+      // Fees, or something else caused the balance to be negative. Just report the change
+      // and quit
+      if (currAmount.eq(new BN(0))) {
+        const targetMintInfo = await getMintInfo(
+          this.provider,
+          isBuy ? tokenBonding.targetMint : tokenBonding.baseMint
+        );
+        return {
+          targetAmount:
+            toNumber(postBalance!, targetMintInfo) -
+            toNumber(preBalance, targetMintInfo),
+        };
+      }
     }
-    const targetMintInfo = await getMintInfo(this.provider, targetMint);
 
+    const targetMintInfo = await getMintInfo(this.provider, targetMint);
     return {
-      targetAmount: toNumber(baseAmount, targetMintInfo),
+      targetAmount: toNumber(currAmount, targetMintInfo),
     };
   }
 
@@ -1844,21 +1914,14 @@ export class SplTokenBonding extends AnchorSdk<SplTokenBondingIDL> {
     const parentKey = (
       await SplTokenBonding.tokenBondingKey(tokenBonding.baseMint)
     )[0];
-    const ret = {
+    const ret = new BondingHierarchy({
       parent: stopAtMint?.equals(tokenBonding.baseMint)
         ? undefined
         : await this.getBondingHierarchy(parentKey, stopAtMint),
       tokenBonding,
       pricingCurve,
-    };
+    });
     (ret.parent || ({} as any)).child = ret;
     return ret;
   }
 }
-
-export type BondingHierarchy = {
-  parent?: BondingHierarchy;
-  child?: BondingHierarchy;
-  tokenBonding: ITokenBonding;
-  pricingCurve: IPricingCurve;
-};